Replace existing credit card demo with latest and rename CKKS_credit_card_fraud
