+++ samples/CKKS_credit_card_fraud/.gitignore
@@ -1,0 +1,4 @@
+CMakeCache.txt
+CMakeFiles
+cmake_install.cmake
+Makefile
