--- samples/credit_card_fraud/.gitignore
@@ -1,4 +1,0 @@
-CMakeCache.txt
-CMakeFiles
-cmake_install.cmake
-Makefile
